pyproject.toml
[tool.poetry]
name = "twitter_plays_pyboy"
version = "0.1.0"
description = ""
authors = ["Kyle Anthony Williams <kyle.anthony.williams2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyboy = "^1.3.0"
Pillow = "^8.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.790"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
